added support for https://github.com/osm0sis/PlayIntegrityFork

+++ Fp_Downloader.apk to use with play integrity by chiteroman/fp.sh
@@ -11,7 +11,13 @@ pm disable eu.xiaomi.module.inject > /dev/null 2>&1 && echo "The miui eu inject 
 echo
 
 echo "[+] Downloading the pif.json"
-/system/bin/curl -L "https://raw.githubusercontent.com/daboynb/autojson/main/pif.json" -o /data/adb/pif.json > /dev/null 2>&1 || /system/bin/curl -L "https://raw.githubusercontent.com/daboynb/autojson/main/pif.json" -o /data/adb/pif.json
+if [ -f /data/adb/modules/playintegrityfix/migrate.sh ]  
+then
+    /system/bin/curl -L "https://raw.githubusercontent.com/daboynb/autojson/main/pif.json" -o /data/adb/modules/playintegrityfix/custom.pif.json > /dev/null 2>&1 || /system/bin/curl -L "https://raw.githubusercontent.com/daboynb/autojson/main/pif.json" -o /data/adb/modules/playintegrityfix/custom.pif.json
+
+else
+    /system/bin/curl -L "https://raw.githubusercontent.com/daboynb/autojson/main/pif.json" -o /data/adb/pif.json > /dev/null 2>&1 || /system/bin/curl -L "https://raw.githubusercontent.com/daboynb/autojson/main/pif.json" -o /data/adb/pif.json
+fi
 echo
 
 echo "[+] Killing com.google.android.gms"
